split fetch implementation across multiple modules
