+++ src/lib.rs
@@ -4,11 +4,11 @@
 //!
 //! * "neq" - NeqAssign trait
 //! * "pure" - Pure components
-//! * "history" - History tracker
 //! * "future" - Async support for Yew Messages
 //! * "fetch" - Wrapper that holds requests and responses.
 //! * "mrc_irc" - Ergonomic Rc pointers.
 //! * "lrc" - Linked-list Rc pointer.
+//! * "history" - History tracker
 //! * "dsl" - Use functions instead of Yew's `html!` macro.
 
 #[cfg(feature = "dsl")]
